mlPipeline/application/ml_model/requirements.txt
numpy
tensorflow-macos
requests